backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-groq
sqlalchemy
alembic
python-multipart
